# src/AppForSEII2526.Web/OpenAPIs/swagger2.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "AppForMovies.API",
    "description": "This API provides services for renting and purchasing movies",
    "contact": {
      "name": "Software Engineering II Team",
      "email": "isii@on.uclm.es"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/license/mit/"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/Rentals/GetDetailsRental": {
      "get": {
        "tags": [
          "Rentals"
        ],
        "operationId": "GetDetailsRental",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RentalDetailDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RentalDetailDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RentalDetailDTO"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Rentals/CreateRental": {
      "post": {
        "tags": [
          "Rentals"
        ],
        "operationId": "CreateRental",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RentalForCreateDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RentalForCreateDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RentalForCreateDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RentalDetailDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RentalDetailDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RentalDetailDTO"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "RentalDetailDTO": {
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 20,
            "minLength": 2,
            "type": "string",
            "nullable": true
          },
          "surname": {
            "maxLength": 100,
            "minLength": 4,
            "type": "string",
            "nullable": true
          },
          "userName": {
            "maxLength": 100,
            "minLength": 4,
            "type": "string",
            "nullable": true
          },
          "deliveryCarDealer": {
            "maxLength": 50,
            "minLength": 1,
            "type": "string",
            "nullable": true
          },
          "paymentMethod": {
            "$ref": "#/components/schemas/RentalPaymentMethodEnum"
          },
          "startDate": {
            "type": "string",
            "format": "date"
          },
          "endDate": {
            "type": "string",
            "format": "date"
          },
          "rentingDate": {
            "type": "string",
            "format": "date"
          },
          "totalPrice": {
            "maximum": 2000,
            "minimum": 30,
            "type": "number",
            "format": "currency"
          },
          "rentalItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RentalItemDTO"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "RentalForCreateDTO": {
        "type": "object",
        "properties": {
          "userName": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "maxLength": 20,
            "minLength": 2,
            "type": "string",
            "nullable": true
          },
          "surname": {
            "maxLength": 100,
            "minLength": 4,
            "type": "string",
            "nullable": true
          },
          "deliveryCarDealer": {
            "maxLength": 20,
            "minLength": 1,
            "type": "string",
            "nullable": true
          },
          "paymentMethod": {
            "$ref": "#/components/schemas/RentalPaymentMethodEnum"
          },
          "startDate": {
            "type": "string",
            "format": "date"
          },
          "endDate": {
            "type": "string",
            "format": "date"
          },
          "rentingDate": {
            "type": "string",
            "format": "date"
          },
          "totalPrice": {
            "type": "number",
            "format": "float"
          },
          "rentalItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RentalItemDTO"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "RentalItemDTO": {
        "type": "object",
        "properties": {
          "model": {
            "maxLength": 20,
            "minLength": 1,
            "type": "string",
            "nullable": true
          },
          "manufacturer": {
            "maxLength": 20,
            "minLength": 1,
            "type": "string",
            "nullable": true
          },
          "rentingPrice": {
            "type": "number",
            "format": "float"
          },
          "quantity": {
            "maximum": 2147483647,
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "RentalPaymentMethodEnum": {
        "enum": [
          "Visa",
          "GooglePay",
          "PayPal"
        ],
        "type": "string"
      },
...
      "ValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "nullable": true
          }
        },
        "additionalProperties": { }
      }
    }
  }
}